Remove stock_changes_helper.rb
